Edu/requirements.txt
fastapi[all]
uvicorn
peewee
pydantic
passlib[bcrypt]
bcrypt==3.2.0
python-jose
# pip install -r requirements.txt
